move timer to app service
